add a dry run option for live database changes

+++ app/Console/Commands/RemoveDuplicatePermissionsCommand.php
@@ -7,7 +7,8 @@ use Illuminate\Console\Command;
 
 class RemoveDuplicatePermissionsCommand extends Command
 {
-    protected $signature = 'utrs-maintenance:remove-duplicate-permissions';
+    protected $signature = 'utrs-maintenance:remove-duplicate-permissions 
+                            {--dry? : A dry run of the command should be ran}';
     protected $description = 'Clears up legacy duplicate permissions from the database';
 
     public function __construct()
@@ -17,6 +18,11 @@ class RemoveDuplicatePermissionsCommand extends Command
 
     public function handle()
     {
+        if (!isset($this->argument('--dry'))) {
+            if (!$this->confirm('You are about to run a live database change. Are you sure? (Did you test it?)')) {
+                return 1;
+            }
+        }
         $duplicates = Permission::with('user')
             ->selectRaw('wiki, userid, count(*) as count')
             ->groupBy(['wiki', 'userid'])
@@ -27,17 +33,25 @@ class RemoveDuplicatePermissionsCommand extends Command
         $progressBar->display();
 
         foreach ($duplicates as $duplicate) {
-            /** @var Permission $duplicate */
-            $this->info("Removing duplicates for user $duplicate->userid in wiki $duplicate->wiki...");
+            if (!isset($this->argument('--dry'))) {
+                /** @var Permission $duplicate */
+                $this->info("Removing duplicates for user $duplicate->userid in wiki $duplicate->wiki...");
 
-            // delete duplicates. can't use $duplicate->delete() due to it being a groupBy
-            Permission::where('userid', $duplicate->userid)->where('wiki', $duplicate->wiki)->delete();
-            // and load correct permissions from wiki
-            $duplicate->user->queuePermissionChecks();
+                // delete duplicates. can't use $duplicate->delete() due to it being a groupBy
+                Permission::where('userid', $duplicate->userid)->where('wiki', $duplicate->wiki)->delete();
+
+                // and load correct permissions from wiki
+                $duplicate->user->queuePermissionChecks();
+            }
+            else {
+                /** @var Permission $duplicate */
+                $this->info("DRY: Wanting to removing duplicates for user $duplicate->userid in wiki $duplicate->wiki");
+            }
 
             $progressBar->advance();
         }
 
         $progressBar->finish();
+        return 0;
     }
 }
